pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dhwani"
version = "0.1.0"
description = "Phonetic normalization toolkit for Hinglish (Romanized Hindi + English code-mixed text)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Krishna Badikela"}
]
keywords = ["hinglish", "hindi", "ipa", "phonetics", "transliteration", "nlp", "normalization"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.21",
]

[project.optional-dependencies]
models = ["ai4bharat-transliteration>=1.0", "epitran>=1.24"]
full = ["ai4bharat-transliteration>=1.0", "epitran>=1.24", "phonemizer>=3.0"]
dev = ["pytest", "pytest-cov", "ruff"]
demo = ["gradio>=4.0"]

[project.scripts]
dhwani = "dhwani.cli:main"

[project.urls]
Homepage = "https://github.com/Kkoundinyaa/dhwani"
Repository = "https://github.com/Kkoundinyaa/dhwani"

[tool.setuptools.packages.find]
include = ["dhwani*"]

[tool.ruff]
line-length = 100
